core: move command line helper
This commit is a preparation to move binaries to each runtime crate.
