+++ runtime/core/Cargo.toml
@@ -11,6 +11,7 @@ license = "GPL-3.0-or-later"
 publish = false
 
 [dependencies]
+libc = "0.2"
 nix = "0.17"
 glib = "0.15"
 hinawa = "0.7"
