+++ src/bin/snd-fireface-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2021 Takashi Sakamoto
 
-use {fireface_runtime::FfRuntime, snd_firewire_ctl_services::*};
+use {core::cmdline::*, fireface_runtime::FfRuntime};
 
 struct FfServiceCmd;
