+++ src/bin/snd-firewire-motu-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
 
-use {motu_runtime::MotuRuntime, snd_firewire_ctl_services::*};
+use {core::cmdline::*, motu_runtime::MotuRuntime};
 
 struct MotuServiceCmd;
