+++ src/bin/snd-firewire-tascam-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
 
-use {snd_firewire_ctl_services::*, tascam_runtime::TascamRuntime};
+use {core::cmdline::*, tascam_runtime::TascamRuntime};
 
 struct TascamServiceCmd;
