+++ src/bin/snd-oxfw-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
 
-use {oxfw_runtime::OxfwRuntime, snd_firewire_ctl_services::*};
+use {core::cmdline::*, oxfw_runtime::OxfwRuntime};
 
 struct OxfwServiceCmd;
